+++ R/utils-gh.R
@@ -170,3 +170,47 @@ is_gh_repo <- function(owner, repo) {
   tryCatch(gh::gh("GET /repos/{owner}/{repo}", repo = repo, 
                   owner = owner), error = function(e) NULL)
 }
+
+
+
+#' **Get current/default git branch name**
+#' 
+#' @noRd
+
+get_git_branch_name <- function() {
+  
+  is_git()
+  
+  current_branch <- gert::git_branch()
+  
+  if (is.null(current_branch)) {
+    
+    config <- as.data.frame(gert::git_config_global())
+    
+    default_global <- config[which(config$"name" == "init.defaultbranch" &
+                                     config$"level" == "global"),
+                             "value"]
+    
+    if (length(default_global) == 1) {
+      
+      current_branch <- default_global
+    
+    } else {
+      
+      default_system <- config[which(config$"name" == "init.defaultbranch" &
+                                       config$"level" == "system"),
+                               "value"]
+      
+      if (length(default_system) == 0) {
+        
+        current_branch <- "master"  
+        
+      } else {
+        
+        current_branch <- default_system
+      }
+    }
+  }
+  
+  current_branch
+}